neurocare-backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
bcrypt==4.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
anthropic==0.39.0
eventlet==0.35.1
